Add files via upload
The decode.py file was used the decode the first part of the homework assignment. I found that one of the most common three-letter words used in the message was "KYV" so I equated that to be "THE" and found that this required a simple Caesar's Cipher with a shift of 9 letters, so K goes to T, Y goes to H, and V goes to E. I used this key to decode the rest of the message using the sample code that was showed during the encryption team's presentation. This resulted in the decoded message of Katherine Johnson biography. The punctuations were unaffected by the decoder.